update category screen design
