+++ Buyva/app/src/main/java/com/example/buyva/data/model/Category.kt
@@ -1,0 +1,3 @@
+package com.example.buyva.data.model
+
+data class Category(val name: String, val imageResId: Int)
