+++ Buyva/app/src/main/java/com/example/buyva/data/model/Product.kt
@@ -1,0 +1,10 @@
+package com.example.buyva.data.model
+
+data class Product(
+    val id: Int,
+    val brand: String,
+    val model: String,
+    val price: String,
+    val imageResId: Int,
+    val category: String
+)
