+++ Buyva/app/src/main/java/com/example/buyva/features/home/view/ProductSection.kt
@@ -15,11 +15,10 @@ import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import com.example.buyva.data.model.Product
 import com.example.buyva.ui.theme.Cold
 import com.example.buyva.ui.theme.ubuntuMedium
 import kotlinx.coroutines.delay
-import com.example.buyva.R
-import com.example.buyva.features.categories.view.Product
 
 @Composable
 fun ProductSection(products: List<Product>) {
